requirements.txt
aniso8601==1.0.0
Babel==2.0
blinker==1.4
dominate==2.1.12
Flask==0.10.1
Flask-Babel==0.9
Flask-Bootstrap==3.3.5.6
Flask-Login==0.3.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.4
Flask-Security==1.7.4
Flask-SQLAlchemy==2.0
Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
passlib==1.6.5
pytz==2015.4
PyYAML==3.11
six==1.9.0
speaklater==1.3
SQLAlchemy==1.0.8
Werkzeug==0.10.4
WTForms==2.0.2
